src/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synergysphere-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlalchemy>=2.0",
  "psycopg[binary,pool]>=3.2",
  "alembic",
  "pydantic>=2",
  "python-multipart",
  "python-dotenv",
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["app*"]        # <-- only package 'app'
exclude = ["alembic*"]    # <-- ignore alembic folder
